rename proxy to prism
